+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## 16.0.0-11
 
 - Fix tabbing for Firefox
+- Fix pressing Enter behaviour inside a form
 
 ## 16.0.0-10